refactor: change dir structure
