+++ src/explorer/fileExplorerProvider.ts
