+++ src/explorer/fileItem.ts
