+++ src/explorer/strategies/fileSystemStrategy.ts
