+++ src/explorer/strategies/fileSystemStrategyBase.ts
