+++ src/explorer/strategies/flatFileSystemStrategy.ts
