+++ src/explorer/strategies/treeFileSystemStrategy.ts
